+++ internal/protocol/list_compareandretainall.go
@@ -37,7 +37,7 @@ func ListCompareAndRetainAllEncodeRequest(name *string, values []*Data) *ClientM
 	clientMessage.SetMessageType(LIST_COMPAREANDRETAINALL)
 	clientMessage.IsRetryable = false
 	clientMessage.AppendString(name)
-	clientMessage.AppendInt(len(values))
+	clientMessage.AppendInt32(int32(len(values)))
 	for _, valuesItem := range values {
 		clientMessage.AppendData(valuesItem)
 	}
